change path to /advent/board
