+++ src/app/[locale]/advent/+components/AdventOfCodeClient.ts
