+++ src/app/[locale]/advent/+components/AdventOfCodeLoader.tsx
@@ -1,0 +1,18 @@
+'use client';
+import { useEffect } from 'react';
+
+interface Props {
+  children: React.ReactNode;
+}
+
+export function AdventOfCodeLoader({ children }: Props) {
+  useEffect(() => {
+    document.body.classList.add('aoc-bg');
+
+    return () => {
+      document.body.classList.remove('aoc-bg');
+    };
+  }, []);
+
+  return children;
+}
