+++ src/app/[locale]/advent/board/+components/Stars.tsx
@@ -2,8 +2,7 @@ import { range } from '@fullstacksjs/toolbox';
 
 import { cn } from '@/utils/cn';
 
-import StarActive from './Star-Active.svg';
-import StarInactive from './Star-Inactive.svg';
+import Star from './Star.svg';
 
 interface StarsProps {
   stars: number;
@@ -16,7 +15,15 @@ export function Stars({ stars, className }: Readonly<StarsProps>) {
       className={cn('items-center text-xs font-semibold text-fg-1', className)}
     >
       {range(25).map((i) => (
-        <span key={i}>{stars >= i ? <StarActive /> : <StarInactive />}</span>
+        <span
+          key={i}
+          className={cn({
+            'text-accent-0': stars >= i,
+            'text-bg-muted': stars < i,
+          })}
+        >
+          <Star />
+        </span>
       ))}
       <span className="ml-3 w-[40px]">120 pt</span>
     </p>
