+++ src/app/[locale]/advent/board/+components/UserTable.tsx
@@ -1,6 +1,6 @@
 import * as Table from '@/components/Table';
 
-import type { AdventOfCodeUser } from './AdventOfCodeClient';
+import type { AdventOfCodeUser } from '../../+components/AdventOfCodeClient';
 import { UserRow } from './UserRow';
 
 interface Props {
@@ -15,9 +15,9 @@ export const UserTable = ({ users }: Props) => {
           <Table.Head className="sticky top-0 px-8 tablet:px-14">
             Rank
           </Table.Head>
-          <Table.Head className="sticky top-0 ">User</Table.Head>
-          <Table.Head className="sticky top-0 hidden px-24 tablet:flex">
-            Rank
+          <Table.Head className="sticky top-0">User</Table.Head>
+          <Table.Head className="sticky top-0 hidden tablet:table-cell">
+            Score
           </Table.Head>
         </Table.Row>
       </Table.Header>
