+++ src/app/[locale]/advent/page.tsx
@@ -1,10 +1,13 @@
-import { NextIntlClientProvider } from 'next-intl';
+import { NextIntlClientProvider, useTranslations } from 'next-intl';
 import { getMessages } from 'next-intl/server';
 import { pick } from 'radash';
 
+import { Article } from '@/components/Article';
+import { Paragraph } from '@/components/Paragraph';
 import { generatePageOG } from '@/components/SEO';
 
-import AdventContent from './AdventContent';
+import { HowToJoin } from './+components/HowToJoin';
+import { i18nMap } from './+components/i18nMap';
 
 export const metadata = generatePageOG({
   title: 'Advent of Code: FullstacksJS',
@@ -14,10 +17,33 @@ export const metadata = generatePageOG({
 
 export default async function AdventOfCodePage() {
   const messages = await getMessages();
+  const t = useTranslations('advent');
+  const title = t.rich('title', i18nMap) as React.ReactElement;
+  const howWorks = t.rich('how-works', i18nMap) as React.ReactElement;
+  const ai = t.rich('ai.title', i18nMap) as React.ReactElement;
 
   return (
     <NextIntlClientProvider messages={pick(messages, ['advent'])}>
-      <AdventContent />
+      <div className="flex flex-col gap-16">
+        <Article id="advent" title={title}>
+          <Paragraph>{t.rich('desc', i18nMap)}</Paragraph>
+          <div>
+            <Paragraph>{t.rich('kick-off', i18nMap)}</Paragraph>
+            <Paragraph>{t.rich('fun', i18nMap)}</Paragraph>
+          </div>
+          <Paragraph>{t.rich('desc-2', i18nMap)}</Paragraph>
+        </Article>
+        <Article id="advent" title={howWorks}>
+          <div>
+            <Paragraph>{t.rich('puzzles', i18nMap)}</Paragraph>
+            <Paragraph>{t.rich('stars', i18nMap)}</Paragraph>
+          </div>
+        </Article>
+        <HowToJoin />
+        <Article id="ai" title={ai}>
+          <Paragraph>{t.rich('ai.desc', i18nMap)}</Paragraph>
+        </Article>
+      </div>
     </NextIntlClientProvider>
   );
 }
